+++ package.json
@@ -12,10 +12,14 @@
     "@react-three/fiber": "^8.11.2",
     "@types/three": "^0.149.0",
     "aframe": "^1.4.1",
+    "localforage": "^1.10.0",
+    "match-sorter": "^6.3.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.10.0",
     "socket.io": "^4.6.1",
     "socket.io-client": "^4.6.1",
+    "sort-by": "^1.2.0",
     "three": "^0.149.0"
   },
   "devDependencies": {
