+++ src/Page/Login.css
@@ -1,0 +1,17 @@
+.Login{
+    height: 100vh;
+    width: 100vw;
+
+    background-color: lightgrey;
+}
+#login_modal{
+    height: 76%;
+    width: 24%;
+    background-color: white;
+
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    border-radius: 5%;
+}
