+++ src/socket/socket.js
@@ -4,9 +4,8 @@
 
 // export { socket };
 
-
 // Websocket
 let socket = new WebSocket("wss://testsocket.danceoftaihou.live");
 // let socket = new WebSocket("ws://localhost:8989")
 
-export {socket};
+export { socket };
